Last refactoring
